+++ EntityClassLibrary/RefitWorkModelGoods.cs
@@ -7,8 +7,8 @@ namespace EntityClassLibrary
 {
     public class RefitWorkModelGoods : BaseEntity 
     {
-        private int count;
-        public virtual int Count
+        private float count;
+        public virtual float Count
         {
             get { return count; }
             set { count = value; }
